Move Button Component up a level
